implement manifest functionality
